requirements.py38.txt
absl-py==1.4.0
aio-pika==8.2.3
aiofiles==23.2.1
aiogram==2.15
aiohappyeyeballs==2.4.3
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiormq==6.4.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
APScheduler==3.9.1.post1
asgiref==3.8.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
Babel==2.14.0
backoff==2.2.1
backports.zoneinfo==0.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
blis==0.7.11
boto3==1.34.91
botocore==1.34.91
build==1.2.1
CacheControl==0.12.14
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==2.2.1
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confection==0.1.4
confluent-kafka==2.3.0
cryptography==42.0.5
cycler==0.12.1
cymem==2.0.8
dask==2022.10.2
dataclasses-json==0.6.4
deepdiff==7.0.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.3.0
docopt==0.6.2
durationpy==0.9
emoji==2.11.1
exceptiongroup==1.2.1
faiss-cpu==1.8.0
fastapi==0.110.2
fbmessenger==6.0.0
filelock==3.13.4
filetype==1.2.0
fire==0.6.0
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.63.0
graphlib_backport==1.1.0
greenlet==3.0.3
grpcio==1.62.2
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
iniconfig==2.0.0
interchange==2021.0.4
itsdangerous==2.2.0
jax==0.4.13
jieba==0.42.1
Jinja2==3.1.3
jiter==0.6.1
jmespath==1.0.1
joblib==1.2.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpickle==3.0.4
jsonpointer==2.4
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.2.16
langchain-community==0.2.17
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langcodes==3.4.0
langdetect==1.0.9
langsmith==0.1.137
language_data==1.2.0
libclang==18.1.1
locket==1.0.0
lxml==5.2.1
marisa-trie==1.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.5.3
mattermostwrapper==2.2
mdurl==0.1.2
ml-dtypes==0.2.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.8
multidict==5.2.0
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==2.6.3
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
onnxruntime==1.17.3
openai==1.52.2
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.10.1
overrides==7.7.0
packaging==20.9
pamqp==3.2.1
pansi==2020.7.3
partd==1.4.1
pillow==10.3.0
pkgutil_resolve_name==1.3.10
pluggy==1.5.0
portalocker==2.8.2
posthog==3.5.0
preshed==3.0.9
prompt-toolkit==3.0.28
propcache==0.2.0
protobuf==4.23.3
psycopg2-binary==2.9.9
py2neo==2021.2.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==1.10.9
pydantic-settings==2.6.0
pydantic_core==2.23.4
pydot==1.4.2
Pygments==2.17.2
PyJWT==2.8.0
pykwalify==1.8.0
pymongo==4.3.3
pyparsing==3.1.2
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pyrsistent==0.20.0
pytest==8.2.0
python-crfsuite==0.9.10
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-iso639==2024.2.7
python-magic==0.4.27
python-socketio==5.11.2
pytz==2022.7.1
PyYAML==6.0.1
questionary==1.10.0
randomname==0.1.5
rapidfuzz==3.8.1
rasa==3.6.20
rasa-sdk==3.6.2
redis==4.6.0
regex==2022.10.31
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rocketchat-API==1.30.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
s3transfer==0.10.1
safetensors==0.4.3
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.10.1
sentry-sdk==1.14.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sklearn-crfsuite==0.3.6
slack_sdk==3.27.1
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.52
srsly==2.4.8
starlette==0.37.2
structlog==23.3.0
structlog-sentry==2.1.0
sympy==1.12
tabulate==0.9.0
tarsafe==0.0.4
tenacity==8.2.3
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-text==2.12.0
termcolor==2.4.0
terminaltables==3.1.10
thinc==8.2.3
threadpoolctl==3.4.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
toolz==0.12.1
tqdm==4.66.2
transformers==4.40.1
twilio==8.2.2
typer==0.9.4
typing-inspect==0.9.0
typing-utils==0.1.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
ujson==5.9.0
unstructured-client==0.22.0
urllib3==1.26.16
uvicorn==0.29.0
uvloop==0.21.0
wasabi==1.1.2
watchfiles==0.21.0
wcwidth==0.2.13
weasel==0.3.4
webexteamssdk==1.6.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.2
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.1
